Changes to update-checkout-config.json (#10)
Moving my local branch changes to update-checkout-config.json into release/5.3-branch-tomswifty.

Regarding changes to update_checkout.py:

I had changed the default scheme from scheme = config['default-branch-scheme'] to scheme = config['release/5.3-branch-tomswifty-branch-scheme'] since I was not seeing the changes occur in the build of binding-tools-for-swift and thought perhaps this was the issue. This change is in this PR https://github.com/xamarin/binding-tools-for-swift-reflector/pull/9/files.

Testing locally, I found that this change was not necessary, and the reason I was not seeing changes in the binding-tools-for-swift build was because I was not updating the hash in binding-tools-for-swift/SwiftVerion.mk so I am reverting back to scheme = config['default-branch-scheme'].

Regarding changes to update-checkout-config.json:

apple's swift-format no longer has a master branch. Thus I am pointing swift-format to apple's swift-5.3-branch branch.
After this PR, I plan to point binding-tools-for-swift/SwiftVersion.mk to this commit's hash, SWIFT_SCHEME=release/5.3, and SWIFT_BRANCH=release/5.3-branch-tomswifty.